Use cppcheck upstream's suggestion for silencing the noise.
When cppcheck cannot find any checkable files it prints this:

cppcheck: error: could not find or open any of the paths given.

In the context of check-all-the-things this isn't useful so it
was requested to be removed but cppcheck upstream rejected that:

https://bugs.debian.org/769757
http://trac.cppcheck.net/ticket/6995

Their recommendation of > null works because cppcheck prints
warnings to stderr instead of stdout. It is also less verbose
when copy-pasting to email and forks one less process.