Add an option to limit the check output lines
This helps make the tool more palatable to more people.

It also avoids licensecheck dumping binary to terminals.

Design decisions:

Default the limit to 10 lines of output.

Offer completions for 0 (all) lines and 1-20 lines.

Use ... to indicate the output of a check was trimmed.

Terminate processes that output too many lines.

Mention the trimmed checks in the other checks list section.