Print spaces when erase to EOL terminal code isn't available
When the TERM variable is set to dumb, el isn't available.

Printing spaces up to the width of the terminal effectively erases the
line but does mean that users selecting the line will get trailing spaces
but that isn't a major issue and can be worked around by users easily.

Reported-by: Trent W. Buck (twb)
Reported-in: #debian-mentors

Traceback (most recent call last):
  File "check-all-the-things", line 1121, in <module>
    main()
  File "check-all-the-things", line 1080, in main
    output = check.do(name, jobs, types, run, hide, limit, method, terminal, remarks)
  File "check-all-the-things", line 601, in do
    output, trimmed = spawn(terminal, method, cmd, hide, header, footer, limit)
  File "check-all-the-things", line 282, in spawn
    show_progress(cmd)
  File "check-all-the-things", line 132, in show_progress
    erase_to_eol_cr()
  File "check-all-the-things", line 120, in erase_to_eol_cr
    sys.stdout.buffer.write(erase_line)
TypeError: a bytes-like object is required, not 'NoneType'