Add flag run-in-tmp-dir for running commands in temporary directories
This is useful for commands that are dangerous when run in untrusted dirs
and for commands that create or modify files in the current directory.

Pass the search directories to find as absolute paths.

Offer {pd} and {cwd} for commands that use search directories.

Run commands in newly created temporary dirs for maximum safety.

Run the commands in subshells to easily delete the temporary dirs on exit.

Prevent running the commands when the directories cannot be entered so
that the commands are not executed in the potentially unsafe current dir.

Suggested-by: Jakub Wilk <jwilk@debian.org>
Suggested-in: <20160627201336.GA9003@jwilk.net>
Reviewed-by: Jakub Wilk <jwilk@jwilk.net>