Add flag run-in-root-dir for running commands in root directory
This is useful for commands that are dangerous when run in untrusted dirs
and for commands that create or modify files in the current directory but
behave fine when run from the root directory.

Run the commands in subshells to avoid a manual cd before exit.

Suggested-by: Jakub Wilk <jwilk@debian.org>
Suggested-in: <20160627201336.GA9003@jwilk.net>